Rename 'module catalog'->'lineage'.
"Catalog" is now the name for the aggregate of Lineages (and all the
other data that should be stored in that tree, such as Replays);
"Lineage" is the thing that has a list of Releases.

Docs site already updated to test out this new naming; it seems
significantly clearer and does a better job of explaining itself, and
has gotten good feedback from proofreaders.

It also, as we can see in a couple places in this diff already,
reduces quite a few incidences of irritating name ambiguity internally.

This'll break a few of the API functions used by reach (nee stellar),
but shouldn't affect much else, since nothing else in the stack directly
interacts with the Layer 2 concepts like Lineages.

Signed-off-by: Eric Myhre <hash@exultant.us>