Drop RunOperation from the repeatr package.
It will reappear in the Reach codebase, under `./gadgets/operation`,
split into two separate functions: `Resolve` and `Eval`.

Having this RunOperation function in the repeatr package here has been
odd since the start; it's clearly all layer-2 business, not layer-1;
and for that matter, in retrospec, it's utterly unclear to me why I
ever even put a function with so much action in it here in the API
repo at all.  So this is housecleaning as much as anything (although
prompted by the desire to split the function for other reasons).

Signed-off-by: Eric Myhre <hash@exultant.us>